move changebyte to parser
