+++ src/entity/borehole_parser.py
@@ -1,16 +1,36 @@
 """Convert boreprofile document to processed entries."""
 
 import logging
+from io import BytesIO
 from pathlib import Path
 
 import pymupdf
 from extraction.runner import extract
+from pymupdf import Document
 
-from src.entity.utils import pages_to_bytes
 from src.page_classes import PageClasses
 from src.page_structure import ProcessedEntities
 
 logger = logging.getLogger(__name__)
+
+
+def pages_to_bytes(pdf_document: Document, page_start: int, page_end: int) -> BytesIO:
+    """Extract a range of pages from a PDF document and return them as a BytesIO buffer.
+
+    Args:
+        pdf_document (Document): PDF to split.
+        page_start (int): Start page (1-based).
+        page_end (int): End page (1-based).
+
+    Returns:
+        BytesIO: Selected subset of pages as bytes.
+    """
+    # Create a new PDF for the selected pages
+    with pymupdf.open() as select_pdf:
+        for page_number in range(page_start, page_end + 1):
+            # Insert the page into the new PDF
+            select_pdf.insert_pdf(pdf_document, from_page=page_number - 1, to_page=page_number - 1)
+        return BytesIO(select_pdf.tobytes())
 
 
 def _find_undetected_pages(
